requirements.txt
aiocache==0.12.2
aiohttp~=3.12.9
APScheduler==3.9.1
Babel~=2.13.1
ballpark @ git+https://github.com/Mee6/python-ballpark.git
chat-exporter==2.8.1
coc.py @ git+https://github.com/mathsman5133/coc.py.git@march-update
dnspython==2.6.1
emoji==2.14.0
expiring-dict==1.1.0
fluent.runtime==0.4.0
hikari==2.3.3
hikari-lightbulb==3.0.0a26
hikari-toolbox==0.1.7
imagehash==4.3.1
ipython==8.13.0
kaleido==0.2.1
loguru==0.7.2
matplotlib==3.10.0
meilisearch-python-sdk==3.4.0
msgspec==0.18.6
num2words @ git+https://github.com/ClashKingInc/num2words.git
numerize==0.12
numpy==1.26.4
openai==0.27.8
openpyxl==3.1.2
orjson==3.10.0
pandas==2.2.3
pendulum==3.0.0
plotly==5.14.1
pymitter==1.0.0
pymongo==4.13.2
python-dateutil==2.8.2
python-dotenv==0.19.2
python-snappy
pytz==2021.3
redis==4.6.0
sentry_sdk==2.19.2
ujson~=5.10.0
uvicorn==0.34.0
Wand==0.6.7
websockets==14.1
XlsxWriter==3.0.3
sentry-sdk~=2.19.2
pillow~=10.4.0
setuptools~=74.1.1
watchdog==6.0.0
wcwidth==0.2.13
